#70 Extracts workstate to its own file, blockstate.go
